requirements.txt
langchain
streamlit
openai
tiktoken
python-dotenv
pinecone-client
pypdf
joblib
pandas
